rednote/requirements.txt
accelerate==1.10.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
attrs==25.3.0
av==15.1.0
blake3==1.0.5
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.11.0
configobj==5.0.9
configparser==7.2.0
datasets==4.0.0
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.3.0
etelemetry==0.3.1
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
filelock==3.19.1
frontend==0.0.3
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
interegular==0.3.3
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
looseversion==1.3.0
lxml==6.0.1
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
mistral-common==1.8.4
mlx==0.29.0
mlx-lm==0.27.1
mlx-metal==0.29.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numba==0.61.2
numpy==2.2.6
openai==1.106.1
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
outlines-core==0.2.10
packaging==25.0
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
pathlib==1.0.1
pillow==11.3.0
pip==25.2
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.2
protobuf==6.32.0
prov==2.1.1
psutil==7.0.0
puremagic==1.30
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydot==4.0.1
pygments==2.19.2
pymupdf==1.26.4
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
pyyaml==6.0.2
pyzmq==27.0.2
qwen-vl-utils==0.0.11
rdflib==7.2.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
safetensors==0.6.2
scipy==1.16.1
sentencepiece==0.2.1
sentry-sdk==2.37.0
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.47.3
sympy==1.14.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traits==7.0.2
transformers==4.56.1
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.2
watchfiles==1.1.0
websockets==15.0.1
xgrammar==0.1.23
xxhash==3.5.0
yarl==1.20.1
